Move send function to utils
